Handle management plan review fail
