+++ submit-api/src/submit_api/enums/item_status.py
@@ -15,3 +15,4 @@ class ItemStatus(enum.Enum):
     COMPLETED = 'COMPLETED'
     SUBMITTED = 'SUBMITTED'
     PASSED_CONSULTATION_CHECK = 'PASSED_CONSULTATION_CHECK'
+    REVIEW_REJECTED = 'REVIEW_REJECTED'
