+++ submit-api/src/submit_api/resources/staff/item.py
@@ -22,6 +22,7 @@ from submit_api.resources.apihelper import Api as ApiHelper
 from submit_api.schemas.item import StaffItemSchema
 from submit_api.schemas.submission_review import SaveSubmissionReviewRequestSchema, SubmissionReviewSchema
 from submit_api.services.item import ItemService
+from submit_api.services.submission_review import SubmissionReviewService
 from submit_api.utils.roles import EpicSubmitRole
 from submit_api.utils.util import cors_preflight
 
@@ -74,5 +75,5 @@ class ItemReview(Resource):
     def post(item_id):
         """Save submission review."""
         request_body = SaveSubmissionReviewRequestSchema().load(API.payload)
-        review = ItemService.save_submission_review(item_id, request_body)
+        review = SubmissionReviewService.save_submission_review(item_id, request_body)
         return SubmissionReviewSchema().dump(review), HTTPStatus.OK
